Split test_name_error.rb
